// src/common/enums/user-module.enum.ts
export enum UserModule {
  DASHBOARD = 'DASHBOARD',
  CUSTOMERS = 'CUSTOMERS',
  SUPPLIERS = 'SUPPLIERS',
  INVENTORY = 'INVENTORY',
  LABORATORIES = 'LABORATORIES',
  BATCHES = 'BATCHES',
  WAREHOUSE = 'WAREHOUSE',
  RECEPTION = 'RECEPTION',
  MOVEMENTS = 'MOVEMENTS',
  PUNTO_VENTA = 'PUNTO_VENTA',
  INVOICES = 'INVOICES',
  BILLING = 'BILLING',
  TICKETS = 'TICKETS',
  REPORTS = 'REPORTS',
  DISCOUNTS = 'DISCOUNTS',
  CATEGORIES = 'CATEGORIES',
  DOCUMENTS = 'DOCUMENTS',
  SETTINGS = 'SETTINGS',
}